[chore] rename actions.py => driver.py
